requirements.txt
opencv-contrib-python
pyinstaller
pyyaml
flask